# added modal to wishlist Revert "modal implementing ... v2" This reverts commit 702fad84f93f285bcd9d79dba348452b1cfe59dc.
