+++ src/components/modal/modal.styles.js
@@ -10,16 +10,21 @@ export const useStyles = makeStyles((theme) => ({
     backgroundColor: theme.palette.background.paper,
     border: '2px solid #000',
     boxShadow: theme.shadows[5],
-    padding: theme.spacing(2, 4, 3)
+    padding: theme.spacing(2, 4, 3),
+    '& p': {
+      fontSize: '1.2em'
+    }
   },
   buttonGroup: {
     width: '100%',
     textAlign: 'center',
     '& > button': {
-      color: theme.palette.button.normal.textColor,
+      color: theme.palette.button.normal.color,
       backgroundColor: theme.palette.button.normal.backgroundColor,
+      padding: 5,
+      margin: 5,
       '&:hover': {
-        color: theme.palette.button.hover.textColor,
+        color: theme.palette.button.hover.color,
         backgroundColor: theme.palette.button.hover.backgroundColor
       }
     }
